requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
confluent-kafka>=2.3.0
pydantic>=2.5.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
requests>=2.31.0
